working-protocols

+++ src/physion/intrinsic/build_protocols.py
@@ -13,7 +13,11 @@ def to_num(direction):
 def build_json(direction='Up', 
                screen='Mouse-Goggles',
                bg_color=0.2,
-               period=12):
+               period=12,
+               center=0,
+               size=7,
+               flicker_size=5,
+               length=200):
 
 
     return """
@@ -30,14 +34,20 @@ def build_json(direction='Up',
     "presentation-interstim-period": 0,
     "N-repeat": 1,
     "direction": %(direction)s,
-    "bar-size": 7,
+    "bar-center": %(center)s,
+    "bar-length": %(length)s,
+    "bar-size": %(size)s,
     "bg-color": %(bg_color)s,
-    "flicker-size": 5,
+    "flicker-size": %(flicker_size)s,
     "flicker-freq": 5
     } 
 """ % {'direction':to_num(direction),
        'period':period,
        'bg_color':bg_color,
+       'size':size,
+       'flicker_size':flicker_size,
+       'center':center,
+       'length':length,
        'screen':screen}
 
 if __name__=='__main__':
@@ -46,6 +56,12 @@ if __name__=='__main__':
     parser=argparse.ArgumentParser(description="Experiment interface",
                        formatter_class=argparse.RawTextHelpFormatter)
 
+    parser.add_argument("protocol", type=str,
+                        help="""
+                        either:
+                            - intrinsic
+                            - ocular-dominance
+                        """)
     parser.add_argument("screen", type=str,
                         help="""
                         either:
@@ -56,23 +72,61 @@ if __name__=='__main__':
     
     args = parser.parse_args()
 
-    for period in [6, 12]:
-        if not os.path.isdir('physion/acquisition/protocols/movies/intrinsic/'):
-            os.mkdir('physion/acquisition/protocols/movies/intrinsic')
-        for direction in ['up', 'down', 'left', 'right']:
-            # create the directory
-            if not os.path.isdir('physion/acquisition/protocols/movies/intrinsic/flickering-bars-period%is' % period):
-                os.mkdir('physion/acquisition/protocols/movies/intrinsic/flickering-bars-period%is' % period)
-            # write the json
-            with open('temp.json', 'w') as f:
-                f.write(build_json(direction, 
-                                   screen=args.screen,
-                                   period=period, 
-                                   bg_color=args.bg_color))
-            # build the movie
-            os.system('python -m physion.visual_stim.build temp.json --wmv')
-            os.rename('movies/temp/movie.wmv', 
-    'physion/acquisition/protocols/movies/intrinsic/flickering-bars-period%is/%s.wmv' % (period, direction))
+    folder = os.path.join('physion', 'acquisition', 'protocols', 'movies', args.protocol)
+
+    if args.protocol=='intrinsic':
+        for period in [6, 12]:
+            if not os.path.isdir(folder):
+                os.mkdir(folder)
+            for direction in ['up', 'down', 'left', 'right']:
+                # create the directory
+                if not os.path.isdir(os.path.join(folder, 'flickering-bars-period%is' % period)):
+                    os.mkdir(os.path.join(folder, 'flickering-bars-period%is' % period))
+                # write the json
+                with open('temp.json', 'w') as f:
+                    f.write(build_json(direction, 
+                                       screen=args.screen,
+                                       period=period, 
+                                       bg_color=args.bg_color))
+                # build the movie
+                os.system('python -m physion.visual_stim.build temp.json --wmv')
+                os.rename(os.path.join('movies', 'temp', 'movie.wmv'),
+                          os.path.join(folder,
+                                       'flickering-bars-period%is' % period,
+                                       '%s.wmv' % direction))
+
+    elif args.protocol=='ocular-dominance':
+        for period in [6, 12]:
+            if not os.path.isdir(folder):
+                os.mkdir(folder)
+            for side in ['left', 'right']:
+                for direction in ['up', 'down']:
+                    # create the directory
+                    if not os.path.isdir(os.path.join(folder, 'flickering-bars-period%is' % period)):
+                        os.mkdir(os.path.join(folder, 'flickering-bars-period%is' % period))
+                    # write the json
+                    with open('temp.json', 'w') as f:
+                        f.write(build_json(direction, 
+                                           screen=args.screen,
+                                           period=period, 
+                                           center=-5 if side=='left' else 5,
+                                           length=20,
+                                           size=2,
+                                           flicker_size=2,
+                                           bg_color=args.bg_color))
+                    # build the movie
+                    """
+                    os.system('python -m physion.visual_stim.build temp.json --wmv')
+                    os.rename(os.path.join('movies', 'temp', 'movie.wmv'),
+                          os.path.join(folder,
+                                       'flickering-bars-period%is' % period,
+                                       '%s-%s.wmv' % (side, direction)))
+                    """
+                    os.system('python -m physion.visual_stim.build temp.json')
+                    os.rename(os.path.join('movies', 'temp', 'movie.mp4'),
+                          os.path.join(folder,
+                                       'flickering-bars-period%is' % period,
+                                       '%s-%s.mp4' % (side, direction)))
 
         os.remove('temp.json')
     shutil.rmtree('movies')
